Errors + Metrics
I've added code to convert the errors and metrics.  There was an issue reading the "parent" attribute of the metric since some of those values are filled with nonValues.  Since the errors and metrics are represented as a tuple containing lists, I was having a hard time figuring out how to pull the item name for each attribute from the JSON.